requirements_dev.txt
autoflake==2.3.1
pytest==8.1.1
isort==5.13.2
mypy==1.9.0
wheel==0.43.0
twine==5.0.0 
invoke==2.2.0
black==24.4.0
flake8==7.0.0
blacken-docs==1.16.0
setuptools==71.1.0
pandas-stubs